Implemented third tier of player states
Set up a third tier of states for the player state machine.
- Attack
- Ranged
- Teleport
- Pull
- Heal
- Empty

The Empty state is used to set up the third tier and each state at this level will derive from it and return to the empty state.

Ranged state has been implemented firing a bolt from the players hand. the bolt goes forward for 4 seconds or until it hits an object.

Implemented a basic enemy that just takes damage until destroyed. Will be expanded when the enemy state machine is implemented.