Light Manager Set Up
Set up the lighting manager script to work for all lights in the scene.
done to begin work on player abilities that require lighting data.

-Improvements for future versions:
Have the light manager work on a range based on player location.
remove enemies and lights from the list when out of a general range (30-40 units?)
This should be done to make the script more efficient in the future.