Lighting changes, UI and Player ability changes
Implemented the Shadow pull ability for the player. Pressing E then clicking on an enemy in range will allow the player to choose a new location for them to be placed. Added a Marker UI element which is connected to both the teleport and Shadow pull abilities. This is done to tell the player where the target will be teleported to.

Modified the Light Manager to work for Player teleportation and using the shadow pull ability on Enemies. The light manager will now check if the Teleport Marker is in light or not and allow the player to complete the action if in shadow.

Added UI above Enemies to display current health
Added Player health and Mana UI elements to the players UI Canvas.